requirements.txt
aiogithubapi
aiohttp
PyJWT
cryptography